refactor: centralize PTY finalization and child lifecycle state
